Node map CBS
I did some work on the CBS solution on the graph map but didn't finish it... the independent solver obviously works but It's not what I want.

I'm going to continue tomorrow with this. I should get it working soon and if not, I can try to just use the student solutions and implement deviations in for those.